requirements.txt
pip~=23.2.1
typing_extensions~=4.9.0
pytest~=7.4.3
pluggy~=1.3.0
iniconfig~=2.0.0
numpy~=1.26.3
certifi~=2024.8.30
urllib3~=2.1.0
idna~=3.4
charset-normalizer~=3.3.2
packaging~=23.2
pyee~=13.0.0
typing_extensions~=4.9.0
greenlet~=3.2.1
playwright~=1.52.0
python-dotenv~=1.1.0
typing_extensions~=4.9.0
click~=8.1.7
Faker~=37.1.0